upd(bot): replace new_inc command with create_inc
